% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppmx.R
\name{gcd}
\alias{gcd}
\title{gendata}
\usage{
gcd(
  n_obs = 100,
  concov = 2,
  K = 2,
  similarity = 1,
  simparm = 1,
  alpha = 1,
  m0 = 0,
  s20 = 1,
  v = 2,
  k0 = 10,
  v0 = 1,
  plot = F
)
}
\description{
gendata
}
