% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genmech}
\alias{genmech}
\title{genmech}
\usage{
genmech(
  npred = 10,
  progscen = 1,
  predscen = 1,
  nnoise = 15,
  nset = 30,
  save = FALSE,
  filename = "myscenario"
)
}
\arguments{
\item{npred}{number of \eqn{Q} predictive covariates used to generate the outcome}

\item{progscen}{Prognostic covariates option:
0 - No prognostic biomarkers;
1 - Prognostic biomarkers are considered in the original scale (default);
2 - Prognostic biomarkers are transformed}

\item{predscen}{Predictive covariates option:
1 - \code{npred} Predictive biomarkers are considered to generate outcomes (default);
2 - \code{nnoise} noisy std normals are added to the design matrix of predictive
covariates in addition to \code{npred} predictive biomarkers considered to
generate outcomes}

\item{nnoise}{number of noisy covariates added to predictive biomarkers}

\item{nset}{number of replicated scenarios generated}

\item{save}{logical. if TRUE the function save the results in a \code{.rda} file.
Default is FALSE}

\item{filename}{Name given to the file is results are saved in \code{.rda} file
Default is \code{myscenario.rda}}
}
\value{
a list of 5 elements.
\itemize{
\item \code{Y}: \eqn{n\times K \times}\code{nset} array. It contains \code{nset} matrices that store the outcome in the form of a multinomial experiment
\item \code{Yord}: \eqn{n\times}\code{nset} matrix. It contains the ordinal outcome for each replicated dataset
\item \code{treatment}: \eqn{n-}dimensional vector. It contains the treatment assigned
\item \code{cov}: \eqn{n\times (Q+2)} matrix. It contains all the biomarkers. The last two columns are the two predictive biomarkers
\item \code{prob}: List of \eqn{T}. Each element is a \eqn{n\times K} matrix containing the response probabilities
}
}
\description{
Generates the data for the simulations scenarios 1a and 1b reported in the paper.
It follows the strategy designed by Ma et al. (2019).
In particular, the outcome is a categorical variable representing \eqn{K=3} benefit-increasing levels.
Patients (\eqn{n=152}) are assigned to \eqn{T=2} competing treatments.
}
\references{
Ma, J., Stingo, F. C., & Hobbs, B. P. (2016). Bayesian predictive modeling
for genomic based personalized treatment selection. \emph{Biometrics},
\strong{72}(2), 575-583.
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/biom.12448}

Ma, J., Stingo, F. C., & Hobbs, B. P. (2019). Bayesian personalized
treatment selection strategies that integrate predictive with prognostic
determinants. \emph{Biometrical Journal}, \strong{61}(4), 902-917.
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/bimj.201700323}
}
