% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genmech_alt}
\alias{genmech_alt}
\title{genmech alternative}
\usage{
genmech_alt(npred = 10, nset = 30)
}
\arguments{
\item{npred}{number of predictive covariates used to generate the outcome}

\item{nset}{number of replicated scenarios generated}
}
\description{
genmech alternative
}
