% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genmech_clu3}
\alias{genmech_clu3}
\title{genmech clustering 3}
\usage{
genmech_clu3(npred = 10, nset = 1)
}
\arguments{
\item{npred}{number of predictive covariates used to generate the outcome}

\item{nset}{number of replicated scenarios generated}
}
\description{
genmech clustering 3
}
