% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genoutcome}
\alias{genoutcome}
\title{genoutcome}
\usage{
genoutcome(nobs, alpha, beta1, beta2, beta3, metx, x2, x3)
}
\arguments{
\item{nobs}{number of observations}

\item{alpha}{intercept}

\item{beta1}{coefficient for the predictive covariates. More precisely it is
a coefficient for the combination of the pca to generate the exponetial.}

\item{beta2}{coefficient for the first prognostic covariate (x2)}

\item{beta3}{coefficient for the second prognostic covariate (x3)}

\item{metx}{combination of the pca to generate the exponetial}

\item{x2}{first prognostic covariate (x2)}

\item{x3}{second prognostic covariate (x3)}
}
\value{
a \eqn{\code{nobs}\times 4} matrix. The first column contains the generated
outcome variable, the other three the probabilities for the categories.
}
\description{
Generates outcome variables. The response is an ordinal outcome variable.
}
\details{
Ma, J., Stingo, F. C., & Hobbs, B. P. (2019). Bayesian personalized
treatment selection strategies that integrate predictive with prognostic
determinants. \emph{Biometrical Journal}, \strong{61}(4), 902-917.
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/bimj.201700323}
}
