% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{getdata}
\alias{getdata}
\title{getdata}
\usage{
getdata(...)
}
\arguments{
\item{...}{literal character strings or names}
}
\description{
It is just an helper function.
Using load() or data() I had to call the dataframe by the name of the object.
It caused the devtools::check() to return this note:
checking R code for possible problems ... NOTE
genmech: no visible binding for global variable ‘simupats’
Undefined global functions or variables:
simupats
}
\details{
It is just a shortcut to get things done. I could have worked on the Lazyload
See https://stackoverflow.com/questions/30951204/load-dataset-from-r-package-using-data-assign-it-directly-to-a-variable
}
