% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsimconNNIG}
\alias{gsimconNNIG}
\title{Similarity function}
\usage{
gsimconNNIG(m0, k0, nu0, s20, sumx, sumx2, n, DD, logout)
}
\arguments{
\item{m0}{location parameter}

\item{k0}{shape gamma parameter}

\item{nu0}{scale gamma parameter}

\item{s20}{variance parameter}

\item{sumx}{cluster sum}

\item{sumx2}{cluster squared sum}

\item{n}{number of observation in current cluster}

\item{DD}{logical if 1 Double Dipper similarity is employed}

\item{logout}{logical if 1 return log value}
}
\description{
Normal-normal-IG Similarity function with x following normal and m,v a normal-IG.
}
