% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_ppmx_ct.R
\name{my_dm_ppmx_ct}
\alias{my_dm_ppmx_ct}
\title{PPMx}
\usage{
my_dm_ppmx_ct(
  y,
  X = NULL,
  Xpred = NULL,
  z = NULL,
  zpred = NULL,
  asstreat = NULL,
  alpha = 1,
  CC = 3,
  reuse = 1,
  PPMx = 1,
  similarity = 1,
  consim = 1,
  gowtot = 1,
  alphagow = 1,
  calibration = 0,
  coardegree = 1,
  similparam,
  modelpriors,
  update_hierarchy = 1,
  hsp = 1,
  iter = 1100,
  burn = 100,
  thin = 1,
  mhtunepar = c(0.05, 0.05),
  nclu_init = 5
)
}
\arguments{
\item{y}{response variable}

\item{X}{covariates}

\item{Xpred}{covariate for prediction}

\item{z}{prognostic covariates}

\item{zpred}{prognostic covariates for prediction}

\item{asstreat}{treatment for patients in training set. vector of charachters}

\item{alpha}{value of \eqn{\alpha} for cohesion function (concentration parameter in DP)}

\item{CC}{number of auxiliary parameters}

\item{reuse}{option for the reuse algorithm by Favaro&Teh. integer 0 or 1.
0 - reuse algorithm is NOT adopted
1 - reuse algorithm is adopted}

\item{similarity}{type of similarity function that is employed for the PPMx prior on partitions. Options are
1 - Auxiliary similarity
2 - Double dipper similarity
3 - Gower dissimilarity}

\item{consim}{integer 1 or 2.  1 implies sim for con var is N-N.  2 implies sim is N-NIG}

\item{gowtot}{if similarity parameter is 3 then Gower dissimilarity is employed. default total.
if gowtot ==0,  then Average Gower dissimilarity is taken}

\item{alphagow}{$\alpha$ parameter to compute gower dissimilarity}

\item{calibration}{whether the similarity should be calibrated. Options are
0 - no calibration
1 - standardize similarity value for each covariate
2 - coarsening is applied so that each similarity is raised to the 1/p power}

\item{coardegree}{option to temper the coarsening
1 - $g(x^\star)^{1/p}$
2 - $g(x^\star)^{1/p^{1/2}}$
3 - $g(x^\star)^{1/p^{1/3}}$}

\item{similparam}{vector containing similarity functions paramaters}

\item{modelpriors}{vector containing prior values for model}

\item{update_hierarchy}{should hypeparameter for BNP intercept be updated? if 1 yes (default)}

\item{iter}{MCMC number of iteration}

\item{burn}{MCMC iteration discarded due to burnin}

\item{thin}{thinning for MCMC}

\item{nobs}{length of response vector}

\item{ncon}{number of continuous covariates}

\item{ncat}{number of categorical covariates}

\item{catvec}{\eqn{ncat \times 1} vector indicating the number of categories for each categorical covariate}

\item{vector}{(dimension \eqn{nobs\times 1}) containing response vector}

\item{xcon}{\eqn{nobs \times ncon} contiguous double vector that contains continuous covariate values}

\item{xcat}{\eqn{nobs \times ncat} contiguous int vector that contains categorical covariate values}

\item{initpc:}{partial correlation initialization for prognostic covariates coefficient (default)}
}
\description{
PPMx
}
