% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_ppmx.R
\name{my_mvn_ppmx}
\alias{my_mvn_ppmx}
\title{PPMx}
\usage{
my_mvn_ppmx(
  y,
  X = NULL,
  alpha = 1,
  CC = 3,
  reuse = 1,
  PPMx = 1,
  similarity = 1,
  consim = 1,
  calibration = 0,
  similparam = c(0, 1, 0.1, 1, 2, 0.1, 1),
  modelpriors,
  mhtune = c(0.5, 0.5),
  iter = 1100,
  burn = 100,
  thin = 1
)
}
\arguments{
\item{alpha}{value of \eqn{\alpha} for cohesion function (concentration parameter in DP)}

\item{CC}{number of auxiliary parameters}

\item{reuse}{option for the reuse algorithm by Favaro&Teh. integer 0 or 1.
0 - reuse algorithm is NOT adopted
1 - reuse algorithm is adopted}

\item{similarity}{type of similarity function that is employed for the PPMx prior on partitions. Options are
1 - Auxiliary similarity
2 - Double dipper similarity}

\item{consim}{integer 1 or 2.  1 implies sim for con var is N-N.  2 implies sim is N-NIG}

\item{calibration}{whether the similarity should be calibrated. Options are
0 - no calibration
1 - standardize similarity value for each covariate
2 - coarsening is applied so that each similarity is raised to the 1/p power}

\item{similparam}{vector containing similarity functions paramaters}

\item{modelpriors}{vector containing prior values for model}

\item{mhtune}{vector containing tuning parameters for MCMC updates}

\item{iter}{MCMC number of iteration}

\item{burn}{MCMC iteration discarded due to burnin}

\item{thin}{thinning for MCMC}

\item{nobs}{length of response vector}

\item{ncon}{number of continuous covariates}

\item{ncat}{number of categorical covariates}

\item{catvec}{\eqn{ncat \times 1} vector indicating the number of categories for each categorical covariate}

\item{vector}{(dimension \eqn{nobs\times 1}) containing response vector}

\item{xcon}{\eqn{nobs \times ncon} contiguous double vector that contains continuous covariate values}

\item{xcat}{\eqn{nobs \times ncat} contiguous int vector that contains categorical covariate values}
}
\description{
PPMx
}
