% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countUT.R
\name{npc}
\alias{npc}
\title{npc}
\usage{
npc(output, trtsgn, myoutot)
}
\arguments{
\item{output}{Must be an array storing the probabilities attributed to \code{nobs}
patients to the 3 different benefit levels of the two competing treatments
for \code{nset} replicas of the simulation study.
Its dimension are \code{nobs} x 6 x \code{nset}, where the first three columns are
the probabilities for treatment 1 and the latter
3 columns store the probabilities for treatment 2.}

\item{trtsgn}{treatment assigned to the patients by design.}

\item{myoutot}{List of \code{nset} outcome variables in ordinal notation}
}
\value{
a \code{nobs} vector storing the npc for each patient,
}
\description{
Function to calculate the npc
}
\details{
See Ma et al. (2019) for all the details.
}
