% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rppmx.R
\name{pp_cs_rg}
\alias{pp_cs_rg}
\title{post processing sketch as described in Richardson & Greene (1997)}
\usage{
pp_cs_rg(output, post, nout, dim, refdim = 1)
}
\description{
post processing sketch as described in Richardson & Greene (1997)
}
