% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_ppmx_ct.R
\name{ppmxct}
\alias{ppmxct}
\title{ppmxct}
\usage{
ppmxct(
  y,
  X = NULL,
  Xpred = NULL,
  Z = NULL,
  Zpred = NULL,
  asstreat = NULL,
  PPMx = 1,
  cohesion = 2,
  kappa = c(1, 30, 10, 1),
  sigma = c(0.005, 1, 10),
  similarity = 1,
  consim = 1,
  similparam,
  calibration = 0,
  coardegree = 1,
  modelpriors,
  update_hierarchy = 1,
  hsp = 1,
  iter = 1100,
  burn = 100,
  thin = 1,
  mhtunepar = c(0.05, 0.05),
  CC = 3,
  reuse = 1,
  nclu_init = 5
)
}
\arguments{
\item{y}{\eqn{n_{train} \times K} matrix of ordinal-valued response variable}

\item{X}{\eqn{n_{train} \times Q} dataframe of predictive biomarkers}

\item{Xpred}{\eqn{n_{test} \times Q} dataframe of predictive covariates of new untreated patient}

\item{Z}{\eqn{n_{train} \times P} dataframe of prognostic covariates}

\item{Zpred}{\eqn{n_{test} \times P} dataframe of prognostic covariates of new untreated patient}

\item{asstreat}{\eqn{n_test} vector of integers encoding treatment received by historical patients}

\item{PPMx}{logical. option for the use of product partition model with covariates. (default is yes)}

\item{cohesion}{type of cohesion function that is employed for the PPMx prior on partitions. Options are
1 - DirichletProcess-like cohesion (DP) cohesion
2 - Normalized Generalized Gamma Process (NGG) cohesion}

\item{kappa}{vector of possible values for \eqn{\kappa} for cohesion function (concentration parameter in DP and NGG)}

\item{sigma}{vector of possible value for \eqn{\sigma} parameter in the cohesion function (reinforcement parameter in NGG)}

\item{similarity}{type of similarity function that is employed for the PPMx prior on partitions. Options are
1 - Auxiliary similarity
2 - Double dipper similarity}

\item{consim}{integer 1 or 2.  1 implies sim for con var is NN.  2 implies sim is NNIG}

\item{similparam}{vector containing similarity functions paramaters}

\item{calibration}{If the similarity function is Auxiliary or Double Dipper, the similarity
can be calibrated. Options are
0 - no calibration
1 - standardize similarity value for each covariate
2 - coarsening is applied so that each similarity is raised to the 1/p power}

\item{coardegree}{If the similarity is coarsened, it is possible to temper the coarsening
1 - \eqn{g(x^*)^{1/p}}
2 - \eqn{g(x^*)^{1/p^{1/2}}}}

\item{modelpriors}{vector containing prior values for model}

\item{update_hierarchy}{should hyperparameter for BNP intercept be updated? if 1 yes (default)}

\item{hsp}{parameter for employ horseshoe prior for coefficients for prognostic markers}

\item{iter}{MCMC number of iteration}

\item{burn}{MCMC iteration discarded due to burnin}

\item{thin}{thinning for MCMC}

\item{mhtunepar}{vector containing tuning parameters for MCMC updates}

\item{CC}{number of auxiliary parameters for Algorithm 8 by Neal (2000)}

\item{reuse}{option for the reuse algorithm by Favaro and Teh (2013). integer 0 or 1.
0 - reuse algorithm is not adopted
1 - reuse algorithm is adopted}

\item{nclu_init}{number of cluster used for partial correlation initialization for prognostic covariates coefficient (default)}
}
\value{
a list of 16 elements.
\itemize{
\item \code{label}: List of \eqn{T} matrices. Each element is a \eqn{n_{test}^{a}\times}\code{nout} matrix.
\code{nout} is the number of MCMC iterations after burnin and thinning
The row of each matrix contains the vector of cluster labels of the historical
patient assigned to given treatment
\item \code{asstreat}: Vector of \eqn{n_{test}} treatment assignment
\item \code{num_treat}: Vector of dimension \eqn{T}. Contains the number
of patients assigned to each treatment \eqn{n^a}, for \eqn{a=1, \dots, T}.
\item \code{nclu}: \eqn{T\times}\code{nout} matrix of total number of cluster at each MCMC iteration for each treatment
\item \code{eta}: \code{nout}\eqn{\times T} List of matrices. Each matrix has dimension \eqn{K\times n^a}.
Each element of the list is the linear predictor matrix for treatment \eqn{a} at each MCMC iteration.
\item \code{beta}: Array of dimensions \eqn{P\times K \times}\code{nout}. It contains the matrix of prognostic factors at each MCMC iteration
\item \code{acc_beta}: \eqn{P\times K} matrix containing the counts each prognostic coefficient has been accepted in the MH step. Note that these counts should be divided by \code{iter}
Each element of the list is a list of 2 matrices (one for each treatment).
\item \code{pi_out}: \eqn{n\times K \times}\code{nout} array. It contains the Multinomial parameters for each patients at each MCMC iteration
\item \code{isypred}: \eqn{n\times K \times}\code{nout} array. It contains the outcome in-sample-prediction for each patients at each MCMC iteration
\item \code{WAIC}: scalar. It is the average WAIC
\item \code{lpml}: scalar. It is the average lpml
\item \code{sigmangg}: \eqn{T\times}\code{nout} matrix. It contains the \eqn{\sigma} parameter of NGGP for each treatment at each MCMC iteration
\item \code{kappangg}: \eqn{T\times}\code{nout} matrix. It contains the \eqn{\kappa} parameter of NGGP for each treatment at each MCMC iteration
\item \code{ypred}: \eqn{n_{pred} \times K \times T \times}\code{nout} array. It contains the predicted outcome for each untreated patients, under all the competing for each treatment at each MCMC iteration
\item \code{ypred}: \eqn{n_{pred} \times K \times T \times}\code{nout} array. It contains the Multinomial parameters for each untreated patients, under all the competing for each treatment at each MCMC iteration
}
}
\description{
Function to predict personalized treatment for \eqn{n_{test}} new untreated patients, given their biomarkers.
It leverages response to treatment, \eqn{P} prognostic, \eqn{Q} predictive biomarkers
of \eqn{n_{train}} historical patients.
It accounts for \eqn{K} ordinal response level and \eqn{T} competing treatments.
}
\references{
Page, G. L. and Quintana, F. A. (2016). Calibrating covariate informed
product partition models. \emph{Statistics and Computing}, \strong{28}(5), 1009–1031.

Neal, R. M. (2000). Markov chain sampling methods for Dirichlet process mixture
models. \emph{Journal of Computational and Graphical Statistics}, \strong{9}(2): 249–265.

Favaro, S., Teh, Y. W., et al. (2013). MCMC for normalized random measure mixture
models. \emph{Statistical Science}, \strong{28}(3): 335–359.
}
