% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_ppmx.R
\name{prior_ppmx}
\alias{prior_ppmx}
\title{ppmx prior}
\usage{
prior_ppmx(
  X = NULL,
  PPMx = 1,
  cohesion = 2,
  alpha = 1,
  sigma = 0.2,
  similarity = 1,
  consim = 1,
  similparam,
  calibration = 0,
  coardegree = 1,
  iter = 1100,
  burn = 100,
  thin = 1,
  nclu_init = 5
)
}
\arguments{
\item{X}{covariates}

\item{PPMx}{option for the use of product partition model with covariates. (default is yes)}

\item{cohesion}{type of cohesion function that is employed for the PPMx prior on partitions. Options are
1 - DirichletProcess-like cohesion (DP) cohesion
2 - Normalized Generalized Gamma Process (NGG) cohesion}

\item{alpha}{value of \eqn{\alpha} for cohesion function (concentration parameter in DP)}

\item{sigma}{value of \eqn{\sigma} for cohesion function (reinforcement parameter in NGG)}

\item{similarity}{type of similarity function that is employed for the PPMx prior on partitions. Options are
1 - Auxiliary similarity
2 - Double dipper similarity}

\item{consim}{integer 1 or 2.  1 implies sim for con var is NN.  2 implies sim is NNIG}

\item{similparam}{vector containing similarity functions paramaters}

\item{calibration}{If the similarity function is Auxiliary or Double Dipper, the similarity
can be calibrated. Options are
0 - no calibration
1 - standardize similarity value for each covariate
2 - coarsening is applied so that each similarity is raised to the 1/p power}

\item{coardegree}{If the similarity is coarsened, it is possible to temper the coarsening
1 - \eqn{g(x^*)^{1/p}}
2 - \eqn{g(x^*)^{1/p^{1/2}}}}

\item{iter}{MCMC number of iteration}

\item{burn}{MCMC iteration discarded due to burnin}

\item{thin}{thinning for MCMC}

\item{nclu_init}{number of cluster used for partial correlation initialization for prognostic covariates coefficient (default)}
}
\value{
Number of cluster at each MCMC iteration
}
\description{
ppmx prior
}
