% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simupats_ext}
\alias{simupats_ext}
\title{228 simulated patients data.}
\format{
A data frame containing 228 observation of 53 variables:
\describe{
\item{1-50}{predictive covariates}
\item{51-53}{prognostic covariates}
}
}
\source{
\url{https://www.pnas.org/content/101/12/4164.full?tab=ds}
}
\usage{
simupats_ext
}
\description{
A data frame 53 genes coming from 228 patients. This data were obtained
to emulate the dependence structure arising in sequencing data.
The simulated scenario starts from a well-known dataset of leukemia. The
process to obtained these data from the original Leukemia dataset is extensively
detailed in Ma et al. (2019). It is a small variation of "simupats" dataset,
designed to have larger number of patients.
}
\references{
Ma, J., Stingo, F. C., & Hobbs, B. P. (2019). Bayesian personalized
treatment selection strategies that integrate predictive with prognostic
determinants. \emph{Biometrical Journal}, \strong{61}(4), 902-917.
\url{https://onlinelibrary.wiley.com/doi/full/10.1002/bimj.201700323}

The Leukemia dataset containing gene expression levels for a total of
5,000 genes across 38 patients.


}
\keyword{datasets}
