% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsimupats}
\alias{tsimupats}
\title{152 simulated patients data.}
\format{
A data frame containing 152 observation of 92 variables:
\describe{
\item{1-90}{predictive covariates}
\item{91-92}{prognostic covariates}
}
}
\source{
url original data + aggiungi reference
}
\usage{
tsimupats
}
\description{
A data frame 92 genes coming from 152 patients. This data were obtained
to emulate the dependence structure arising in sequencing data.
The simulated scenario starts from a well-known dataset of leukemia. The
process to obtained these data is extensively detailed iin Ma et al. 2016, 2019.
}
\keyword{datasets}
