% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genmech}
\alias{genmech}
\title{genmech}
\usage{
genmech(
  npred = 10,
  mydata = "data/Simu152pats.txt",
  progscen = 1,
  predscen = 1,
  nnoise = 15,
  nset = 30
)
}
\arguments{
\item{npred}{number of predictive covariates used to generate the outcome}

\item{mydata}{input dataset file}

\item{progscen}{Prognostic covariates option:
1 - Prognostic biomarkers are considered in the original scale (default)
2 - Prognostic biomarkers are transformed}

\item{predscen}{Predictive covariates option:
1 - 10 Predictive biomarkers are considered to generate outcomes (default)
2 - \code{nnoise} noisy std normals are added to the design matrix of predictive
covariates in addition to 10 the predictive biomarkers considered to
generate outcomes}

\item{nnoise}{number of noisy covariates added to predictive biomarkers}

\item{nset}{number of replicated scenarios generated}
}
\value{
a list of 5 elements.
\itemize{
\item 1 - List of \code{nset} outcome variables.
\item 2 - List of \code{nset} outcome variables in ordinal notation.
\item 3 - A vector of assigned treatments.
\item 4 - A matrix of all biomarkers.
\item 5 - A list of outcome probabilities (one element for each treatment).
}
}
\description{
Generates the data for the different simulations scenarios. It follows the
strategy designed by Ma et al. (2019). It is used to generate the data for all
the simulation study scenarios we used. In particular, it has three benefit-incresing levels. Patients are assigned
to two competing treatments.
The ordinal outcome variable are generated using two separate
continuation-ratio logistic functions. The information carried by many
predictive features can be summarized via dimension reduction techniques.
This implementation, in particular, considers thr first two principal
components of a PCA.
}
\details{
See Ma et al. (2019) Supplementary Material for all the details.
}
