% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendata.R
\name{genoutcome}
\alias{genoutcome}
\title{genoutcome}
\usage{
genoutcome(nobs, alpha, beta1, beta2, beta3, metx, x2, x3)
}
\arguments{
\item{nobs}{number of observations}

\item{alpha}{intercept}

\item{beta1}{coefficient for the predictive covariates. More precisely it is
a coefficient for the combination of the pca to generate the exponetial.}

\item{beta2}{coefficient for the first prognostic covariate (x2)}

\item{beta3}{coefficient for the second prognostic covariate (x3)}

\item{metx}{combination of the pca to generate the exponetial}

\item{x2}{first prognostic covariate (x2)}

\item{x3}{second prognostic covariate (x3)}
}
\value{
a \eqn{\code{nobs}\times 4} matrix. The first column contains the generated
outcome variable, the other three the probabilities for the categories.
}
\description{
Generates outcome variables. The response is an ordinal outcome variable.
}
\details{
See Ma et al. (2019) Supplementary Material for all the details.
}
