% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_ppmx_ct.R
\name{ppmxct}
\alias{ppmxct}
\title{ppmxct}
\usage{
ppmxct(
  y,
  X = NULL,
  Xpred = NULL,
  Z = NULL,
  Zpred = NULL,
  asstreat = NULL,
  PPMx = 1,
  cohesion = 2,
  kappa = 1,
  sigma = 0.2,
  similarity = 1,
  consim = 1,
  similparam,
  calibration = 0,
  coardegree = 1,
  modelpriors,
  update_hierarchy = 1,
  hsp = 1,
  iter = 1100,
  burn = 100,
  thin = 1,
  mhtunepar = c(0.05, 0.05),
  CC = 3,
  reuse = 1,
  nclu_init = 5
)
}
\arguments{
\item{y}{ordinal-valued response variable}

\item{X}{predictive biomarkers}

\item{Xpred}{predictive covariates for new untreated patient}

\item{Z}{prognostic covariates}

\item{Zpred}{prognostic covariates for new untreated patient}

\item{asstreat}{treatment for patients in training set. vector of integers}

\item{PPMx}{option for the use of product partition model with covariates. (default is yes)}

\item{cohesion}{type of cohesion function that is employed for the PPMx prior on partitions. Options are
1 - DirichletProcess-like cohesion (DP) cohesion
2 - Normalized Generalized Gamma Process (NGG) cohesion}

\item{kappa}{value of \eqn{\kappa} for cohesion function (concentration parameter in DP)}

\item{sigma}{value of \eqn{\sigma} for cohesion function (reinforcement parameter in NGG)}

\item{similarity}{type of similarity function that is employed for the PPMx prior on partitions. Options are
1 - Auxiliary similarity
2 - Double dipper similarity}

\item{consim}{integer 1 or 2.  1 implies sim for con var is NN.  2 implies sim is NNIG}

\item{similparam}{vector containing similarity functions paramaters}

\item{calibration}{If the similarity function is Auxiliary or Double Dipper, the similarity
can be calibrated. Options are
0 - no calibration
1 - standardize similarity value for each covariate
2 - coarsening is applied so that each similarity is raised to the 1/p power}

\item{coardegree}{If the similarity is coarsened, it is possible to temper the coarsening
1 - \eqn{g(x^*)^{1/p}}
2 - \eqn{g(x^*)^{1/p^{1/2}}}}

\item{modelpriors}{vector containing prior values for model}

\item{update_hierarchy}{should hyperparameter for BNP intercept be updated? if 1 yes (default)}

\item{hsp}{parameter for employ horseshoe prior for coefficients for prognostic markers}

\item{iter}{MCMC number of iteration}

\item{burn}{MCMC iteration discarded due to burnin}

\item{thin}{thinning for MCMC}

\item{mhtunepar}{vector containing tuning parameters for MCMC updates}

\item{CC}{number of auxiliary parameters for algorithm 8 by Neal}

\item{reuse}{option for the reuse algorithm by Favaro&Teh. integer 0 or 1.
0 - reuse algorithm is not adopted
1 - reuse algorithm is adopted}

\item{nclu_init}{number of cluster used for partial correlation initialization for prognostic covariates coefficient (default)}
}
\value{
List
}
\description{
ppmxct
}
