% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatppmx.R
\docType{package}
\name{treatppmx}
\alias{treatppmx}
\title{Treatment PPMx}
\description{
This package implements the method developed by Pedone, Argiento, Stingo to
perform personalized treatment selection for oncological patients.
Both prognostic and predictive biomarkers are taken into accounto to
select the optimal treatment. In particular, redictive biomarkers are
exploited to inform a random partition model to obtain homogeneous clusters.
The implementation has been done in \code{C++} through the use of
\code{Rcpp} and \code{RcppArmadillo}.
}
\references{
TBA
}
\author{
Matteo Pedone

Maintainer: Matteo Pedone \email{matteo.pedone@unifi.it}
}
