% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatppmx.R
\docType{package}
\name{treatppmx}
\alias{treatppmx}
\title{Treatment PPMx}
\description{
The implementation has been done in \code{C++} through the use of \code{Rcpp} and \code{RcppArmadillo}.
}
\author{
matt
}
